move inference output checks
